requirements.txt
Pillow==10.1.0
rasterio==1.3.9
